Move EUPS pkg discovery to pkgdiscovery module
This keeps all the utilities related to finding packages and their
contents in one package.

During this move we've also added type annotations.

discover_setup_packages takes a scope parameter that lets us filter the
package list down to a limited scope (for example, the contents of an
EUPS table file), rather than forcing us to filter later.
