+++ documenteer/stackdocs/build.py
@@ -5,9 +5,10 @@ __all__ = ('build_stack_docs',)
 
 import logging
 import os
-import re
 
-from .pkgdiscovery import find_package_docs, NoPackageDocs
+from .pkgdiscovery import (
+    discover_setup_packages, find_table_file, list_packages_in_eups_table,
+    find_package_docs, NoPackageDocs)
 from ..sphinxrunner import run_sphinx
 
 
@@ -56,24 +57,15 @@ def build_stack_docs(root_project_dir, skippedNames=None):
         logger.info('Creating _static/ at {0}'.format(root_static_dir))
         os.makedirs(root_static_dir)
 
-    # Find package setup by EUPS
-    packages = discover_setup_packages()
-
     # Get packages explicitly required in the table file to filter out
     # implicit dependencies later.
     table_path = find_table_file(root_project_dir)
-    with open(table_path) as fp:
-        table_data = fp.read()
-    listed_packages = list_packages_in_eups_table(table_data)
+    listed_packages = list_packages_in_eups_table(table_path.read_text())
+    # Find package setup by EUPS
+    packages = discover_setup_packages(scope=listed_packages)
 
     # Link to documentation directories of packages from the root project
     for package_name, package_info in packages.items():
-        if package_name not in listed_packages:
-            logger.debug(
-                'Filtering %s from build since it is not explictly '
-                'required by the %s table file.',
-                package_name, table_path)
-            continue
         try:
             package_docs = find_package_docs(
                 package_info['dir'],
@@ -90,96 +82,6 @@ def build_stack_docs(root_project_dir, skippedNames=None):
     # Trigger the Sphinx build
     return_code = run_sphinx(root_project_dir)
     return return_code
-
-
-def discover_setup_packages():
-    """Summarize packages currently set up by EUPS, listing their
-    set up directories and EUPS version names.
-
-    Returns
-    -------
-    packages : `dict`
-       Dictionary with keys that are EUPS package names. Values are
-       dictionaries with fields:
-
-       - ``'dir'``: absolute directory path of the set up package.
-       - ``'version'``: EUPS version string for package.
-
-    Notes
-    -----
-    This function imports the ``eups`` Python package, which is assumed to
-    be available in the build environmen. This function is designed to
-    encapsulate all direct EUPS interactions need by the stack documentation
-    build process.
-    """
-    logger = logging.getLogger(__name__)
-
-    # Not a PyPI dependency; assumed to be available in the build environment.
-    import eups
-
-    eups_client = eups.Eups()
-    products = eups_client.getSetupProducts()
-
-    packages = {}
-    for package in products:
-        name = package.name
-        info = {
-            'dir': package.dir,
-            'version': package.version
-        }
-        packages[name] = info
-        logger.debug('Found setup package: {name} {version} {dir}'.format(
-            name=name, **info))
-
-    return packages
-
-
-def find_table_file(root_project_dir):
-    """Find the EUPS table file for a project.
-
-    Parameters
-    ----------
-    root_project_dir : `str`
-        Path to the root directory of the main documentation project. This
-        is the directory containing the ``conf.py`` file and a ``ups``
-        directory.
-
-    Returns
-    -------
-    table_path : `str`
-        Path to the EUPS table file.
-    """
-    ups_dir_path = os.path.join(root_project_dir, 'ups')
-    table_path = None
-    for name in os.listdir(ups_dir_path):
-        if name.endswith('.table'):
-            table_path = os.path.join(ups_dir_path, name)
-            break
-    if not os.path.exists(table_path):
-        raise RuntimeError(
-            'Could not find the EUPS table file at {}'.format(table_path))
-    return table_path
-
-
-def list_packages_in_eups_table(table_text):
-    """List the names of packages that are required by an EUPS table file.
-
-    Parameters
-    ----------
-    table_text : `str`
-        The text content of an EUPS table file.
-
-    Returns
-    -------
-    names : `list` [`str`]
-        List of package names that are required byy the EUPS table file.
-    """
-    logger = logging.getLogger(__name__)
-    # This pattern matches required product names in EUPS table files.
-    pattern = re.compile(r'setupRequired\((?P<name>\w+)\)')
-    listed_packages = [m.group('name') for m in pattern.finditer(table_text)]
-    logger.debug('Packages listed in the table file: %r', listed_packages)
-    return listed_packages
 
 
 def link_directories(root_dir, package_doc_dirs):
